# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TruelieBot API",
    "version": "1.0.0",
    "description": "API para gesti\u00f3n de conversaciones."
  },
  "paths": {
    "/api/conversations": {
      "get": {
        "summary": "Obtener conversaciones por perfil (con paginaci\u00f3n)",
        "parameters": [
          {
            "name": "profile",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de conversaciones"
          },
          "404": {
            "description": "No se encontraron conversaciones"
          }
        }
      },
      "post": {
        "summary": "Crear una nueva conversaci\u00f3n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "profile": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string"
                  }
                },
                "required": [
                  "profile",
                  "message"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conversaci\u00f3n creada"
          },
          "400": {
            "description": "Datos inv\u00e1lidos"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}